+++ AssetInformationApi/Startup.cs
@@ -42,35 +42,10 @@ using AssetInformationApi.V1.Factories;
 using AssetInformationApi.V1.Infrastructure;
 using Hackney.Core.Middleware;
 using AssetInformationApi.V1.Gateways.Interfaces;
-using Microsoft.AspNetCore.Http;
-using System.Threading.Tasks;
+using AssetInformationApi.V1.Middleware;
 
 namespace AssetInformationApi
 {
-
-    public class TraceLoggingMiddleware
-    {
-        private readonly RequestDelegate _next;
-        private readonly ILogger<TraceLoggingMiddleware> _logger;
-
-        public TraceLoggingMiddleware(RequestDelegate next, ILogger<TraceLoggingMiddleware> logger)
-        {
-            _next = next;
-            _logger = logger;
-        }
-
-        public async Task InvokeAsync(HttpContext context)
-        {
-            // Retrieve the trace ID from the incoming request headers
-            context.Request.Headers.TryGetValue("X-Amzn-Trace-Id", out var traceId);
-
-            _logger.LogInformation($"Incoming Trace ID: {traceId}");
-
-            // Call the next middleware in the pipeline
-            await _next(context);
-        }
-    }
-
     [ExcludeFromCodeCoverage]
     public class Startup
     {
@@ -79,7 +54,6 @@ namespace AssetInformationApi
             Configuration = configuration;
 
             AWSSDKHandler.RegisterXRayForAllServices();
-
             AWSXRayRecorder.InitializeInstance(Configuration);
             AWSXRayRecorder.RegisterLogger(LoggingOptions.SystemDiagnostics);
         }
